Fixed libcommon.a and libxpwn.a building. OpenSSL 1.0.2 installation should be in the next commit.

+++ install.sh
@@ -4,36 +4,39 @@
 # You can find all of the binaries inside their respective folders, for example, "dmg" will be located
 # in the dmg folder. The "ipsw" binary will be located in ipsw-patch and etc. 
 
-# Command to use openssl 1.0.x
 
-cmake -DOPENSSL_ROOT_DIR=/usr/local/opt/openssl@1.0 -DOPENSSL_LIBRARIES=/usr/local/opt/openssl@1.0/lib .
+if [ -e /usr/local/opt/openssl@1.0 ]; then
+	echo -e "Lucky you! You still have the old openssl libraries, like I did before I lost my mind trying to add version checking and such. Building!"
+	cmake -DOPENSSL_ROOT_DIR=/usr/local/opt/openssl@1.0 -DOPENSSL_LIBRARIES=/usr/local/opt/openssl@1.0 .
+else
+	echo -e "OpenSSL will be automatically built/installed soon. Having more issues trying to make this work."
+fi
 
 make clean
 make all
 
-# If libxpwn.a does not exist
-
-if [ ! -e /usr/local/lib/libxpwn.a ]; then
-	make libXPwn.a
-else
-	cp -rnv ipsw-patch/libxpwn.a /usr/local/lib
-fi
-	
-# If libcommon.a does not exist
+# If /usr/local/lib/libcommon.a does not exist
 
 if [ ! -e /usr/local/lib/libcommon.a ]; then
-	make common
-else
-	cp -rnv common/libcommon.a /usr/local/lib
+	# If common/libcommon.a exists
+	if [ -e common/libcommon.a ]; then
+		cp -rnv common/libcommon.a /usr/local/lib
+	fi
+fi
+
+# If /usr/local/lib/libxpwn.a does not exist
+
+if [ ! -e /usr/local/lib/libxpwn.a ]; then
+	# If ipsw-patch/libxpwn.a exists
+	if [ -e ipsw-patch/libxpwn.a ]; then
+		cp -rnv ipsw-patch/libxpwn.a /usr/local/lib
+	fi
 fi
 
 # The modified headers just have the header paths fixed. Replaced "<>" with quotes.
 
 if [ ! -e /usr/local/include/xpwn ]; then
-	echo -e "Installing the modifided headers from includes to /usr/local/include for programs like libipatcher so they can find these headers"
+	echo -e "Installing the modified headers from includes to /usr/local/include for programs like libipatcher so they can find these headers"
 	mkdir -p /usr/local/include/xpwn
 	unzip -d /usr/local/include/xpwn xpwn-modified-headers.zip
-else
-	echo -e "Headers already installed!"
 fi
-
